提交结果

+++ GatherProduct/src/main/java/com/gathermall/product/service/impl/CategoryServiceImpl.java
@@ -1,13 +1,26 @@
 package com.gathermall.product.service.impl;
 
+import com.gathermall.product.entity.CategoryBrandRelation;
 import com.gathermall.product.service.CategoryBrandRelationService;
+import com.gathermall.product.vo.Catelog2Vo;
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
+import org.redisson.api.RLock;
+import org.redisson.api.RReadWriteLock;
+import org.redisson.api.RedissonClient;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.cache.annotation.Caching;
+import org.springframework.data.redis.core.StringRedisTemplate;
+import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
@@ -31,6 +44,13 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryDao, Category> impl
     private CategoryBrandRelationService categoryBrandRelationService;
 
 
+    @Autowired
+    private StringRedisTemplate stringRedisTemplate;
+
+    @Autowired
+    private RedissonClient redissonClient;
+
+
     @Override
     public PageUtils queryPage(Map<String, Object> params) {
         IPage<Category> page = this.page(
@@ -43,7 +63,9 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryDao, Category> impl
 
     @Override
     public List<Category> listWithTree() {
+        // 1 查出所有分类
         List<Category> entities = baseMapper.selectList(null);
+        // 2 组装成父子的树形结构
         List<Category> level1Menus = entities.stream().filter((category) -> {
             return category.getParentCid() == 0;
         }).map((menu) -> {
@@ -58,12 +80,19 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryDao, Category> impl
     @Override
     public Long[] findCatelogPath(Long catelogId) {
         List<Long> paths = new ArrayList<>();
+        //递归查询是否还有父节点
         List<Long> parentPath = findParentPath(catelogId, paths);
-
+        //进行一个逆序排列
         Collections.reverse(parentPath);
         return parentPath.toArray(new Long[parentPath.size()]);
     }
 
+    //同时进行多种缓存操作：@Caching
+    // @Caching(evict = {
+    //         @CacheEvict(value = "category",key = "'getLevel1Categorys'"),
+    //         @CacheEvict(value = "category",key = "'getCatalogJson'")
+    // })
+    @CacheEvict(value = "category", allEntries = true)       //删除某个分区下的所有数据
     @Override
     public void updateCascade(Category category) {
         this.updateById(category);
@@ -74,8 +103,17 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryDao, Category> impl
 
     @Override
     public void removeMenuByIds(List<Long> asList) {
-        //TODO 检查当前删除的菜单id，是否被别的地方引用
-        baseMapper.deleteBatchIds(asList);
+        //检查当前删除的菜单id，是否被别的地方引用
+        List<CategoryBrandRelation> categoryBrandRelation = categoryBrandRelationService
+                .list(new QueryWrapper<CategoryBrandRelation>()
+                        .in("catelog_id", asList));
+
+        if (categoryBrandRelation.size() == 0) {
+            //逻辑删除
+            baseMapper.deleteBatchIds(asList);
+        } else {
+            throw new RuntimeException("该菜单下面还有属性，无法删除!");
+        }
     }
 
 
@@ -114,4 +152,199 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryDao, Category> impl
         }).collect(Collectors.toList());
         return children;
     }
+
+
+    // 每一个需要缓存的数据我们都要指定放到哪个名字的缓存。【缓存的分区】
+    // 当前方法的结果需要缓存，如果缓存中有，方法不调用，如果缓存中没有，会调用方法，最后将方法的结果放入缓存
+    @Cacheable(value = {"category"}, key = "#root.method.name", sync = true)
+    @Override
+    public List<Category> getLevel1Categorys() {
+        System.out.println("getLevel1Categorys........");
+        long l = System.currentTimeMillis();
+        List<Category> categoryEntities = this.baseMapper.selectList(
+                new QueryWrapper<Category>().eq("parent_cid", 0));
+        System.out.println("消耗时间：" + (System.currentTimeMillis() - l));
+        return categoryEntities;
+    }
+
+
+    @Cacheable(value = "category", key = "#root.methodName")
+    @Override
+    public Map<String, List<Catelog2Vo>> getCatalogJson() {
+        System.out.println("查询了数据库");
+
+        //将数据库的多次查询变为一次
+        List<Category> selectList = this.baseMapper.selectList(null);
+
+        //1、查出所有分类
+        //1、1）查出所有一级分类
+        List<Category> level1Categorys = getParent_cid(selectList, 0L);
+
+        //封装数据
+        Map<String, List<Catelog2Vo>> parentCid = level1Categorys.stream().collect(Collectors.toMap(k -> k.getCatId().toString(), v -> {
+            //1、每一个的一级分类,查到这个一级分类的二级分类
+            List<Category> categoryEntities = getParent_cid(selectList, v.getCatId());
+
+            //2、封装上面的结果
+            List<Catelog2Vo> catelog2Vos = null;
+            if (categoryEntities != null) {
+                catelog2Vos = categoryEntities.stream().map(l2 -> {
+                    Catelog2Vo catelog2Vo = new Catelog2Vo(v.getCatId().toString(), null, l2.getCatId().toString(), l2.getName().toString());
+
+                    //1、找当前二级分类的三级分类封装成vo
+                    List<Category> level3Catelog = getParent_cid(selectList, l2.getCatId());
+
+                    if (level3Catelog != null) {
+                        List<Catelog2Vo.Category3Vo> category3Vos = level3Catelog.stream().map(l3 -> {
+                            //2、封装成指定格式
+                            Catelog2Vo.Category3Vo category3Vo = new Catelog2Vo.Category3Vo(l2.getCatId().toString(), l3.getCatId().toString(), l3.getName());
+
+                            return category3Vo;
+                        }).collect(Collectors.toList());
+                        catelog2Vo.setCatalog3List(category3Vos);
+                    }
+
+                    return catelog2Vo;
+                }).collect(Collectors.toList());
+            }
+
+            return catelog2Vos;
+        }));
+
+        return parentCid;
+    }
+
+
+    public Map<String, List<Catelog2Vo>> getCatalogJson2() {
+        //给缓存中放json字符串，拿出的json字符串，反序列为能用的对象
+
+        /**
+         * 1、空结果缓存：解决缓存穿透问题
+         * 2、设置过期时间(加随机值)：解决缓存雪崩
+         * 3、加锁：解决缓存击穿问题
+         */
+
+        //1、加入缓存逻辑,缓存中存的数据是json字符串
+        //JSON跨语言。跨平台兼容。
+        ValueOperations<String, String> ops = stringRedisTemplate.opsForValue();
+        String catalogJson = ops.get("catalogJson");
+        if (StringUtils.isEmpty(catalogJson)) {
+            System.out.println("缓存不命中...查询数据库...");
+            //2、缓存中没有数据，查询数据库
+            Map<String, List<Catelog2Vo>> catalogJsonFromDb = getCatalogJsonFromDbWithRedissonLock();
+
+            return catalogJsonFromDb;
+        }
+
+        System.out.println("缓存命中...直接返回...");
+        //转为指定的对象
+        Map<String, List<Catelog2Vo>> result = new Gson().fromJson(catalogJson, new TypeToken<Map<String, List<Catelog2Vo>>>() {
+        }.getType());
+        //Map<String, List<Catelog2Vo>> result = JSON.parseObject(catalogJson,new TypeReference<Map<String, List<Catelog2Vo>>>(){});
+
+        return result;
+    }
+
+    public Map<String, List<Catelog2Vo>> getCatalogJsonFromDbWithRedissonLock() {
+
+        //1、占分布式锁。去redis占坑
+        //（锁的粒度，越细越快:具体缓存的是某个数据，11号商品） product-11-lock
+        //RLock catalogJsonLock = redissonClient.getLock("catalogJson-lock");
+        //创建读锁
+        RReadWriteLock readWriteLock = redissonClient.getReadWriteLock("catalogJson-lock");
+
+        RLock rLock = readWriteLock.readLock();
+
+        Map<String, List<Catelog2Vo>> dataFromDb = null;
+        try {
+            rLock.lock();
+            //加锁成功...执行业务
+            dataFromDb = getDataFromDb();
+        } finally {
+            rLock.unlock();
+        }
+        //先去redis查询下保证当前的锁是自己的
+        //获取值对比，对比成功删除=原子性 lua脚本解锁
+        // String lockValue = stringRedisTemplate.opsForValue().get("lock");
+        // if (uuid.equals(lockValue)) {
+        //     //删除我自己的锁
+        //     stringRedisTemplate.delete("lock");
+        // }
+
+        return dataFromDb;
+
+    }
+
+
+    private Map<String, List<Catelog2Vo>> getDataFromDb() {
+        //得到锁以后，我们应该再去缓存中确定一次，如果没有才需要继续查询
+        String catalogJson = stringRedisTemplate.opsForValue().get("catalogJson");
+        Gson gson = new Gson();
+        if (!StringUtils.isEmpty(catalogJson)) {
+            //缓存不为空直接返回
+
+            Map<String, List<Catelog2Vo>> result = gson.fromJson(catalogJson, new TypeToken<Map<String, List<Catelog2Vo>>>() {
+            }.getType());
+
+            return result;
+        }
+
+        System.out.println("查询了数据库");
+
+        /**
+         * 将数据库的多次查询变为一次
+         */
+        List<Category> selectList = this.baseMapper.selectList(null);
+
+        //1、查出所有分类
+        //1、1）查出所有一级分类
+        List<Category> level1Categorys = getParent_cid(selectList, 0L);
+
+        //封装数据
+        Map<String, List<Catelog2Vo>> parentCid = level1Categorys.stream().collect(Collectors.toMap(k -> k.getCatId().toString(), v -> {
+            //1、每一个的一级分类,查到这个一级分类的二级分类
+            List<Category> categoryEntities = getParent_cid(selectList, v.getCatId());
+
+            //2、封装上面的结果
+            List<Catelog2Vo> catelog2Vos = null;
+            if (categoryEntities != null) {
+                catelog2Vos = categoryEntities.stream().map(l2 -> {
+                    Catelog2Vo catelog2Vo = new Catelog2Vo(v.getCatId().toString(), null, l2.getCatId().toString(), l2.getName().toString());
+
+                    //1、找当前二级分类的三级分类封装成vo
+                    List<Category> level3Catelog = getParent_cid(selectList, l2.getCatId());
+
+                    if (level3Catelog != null) {
+                        List<Catelog2Vo.Category3Vo> category3Vos = level3Catelog.stream().map(l3 -> {
+                            //2、封装成指定格式
+                            Catelog2Vo.Category3Vo category3Vo = new Catelog2Vo.Category3Vo(l2.getCatId().toString(), l3.getCatId().toString(), l3.getName());
+
+                            return category3Vo;
+                        }).collect(Collectors.toList());
+                        catelog2Vo.setCatalog3List(category3Vos);
+                    }
+
+                    return catelog2Vo;
+                }).collect(Collectors.toList());
+            }
+
+            return catelog2Vos;
+        }));
+
+        //3、将查到的数据放入缓存,将对象转为json
+        String valueJson = gson.toJson(parentCid);
+        stringRedisTemplate.opsForValue().set("catalogJson", valueJson, 1, TimeUnit.DAYS);
+
+        return parentCid;
+    }
+
+
+    private List<Category> getParent_cid(List<Category> selectList, Long parentCid) {
+        List<Category> categoryEntities = selectList.stream().filter(item ->
+                item.getParentCid().equals(parentCid)).collect(Collectors.toList());
+        return categoryEntities;
+        // return this.baseMapper.selectList(
+        //         new QueryWrapper<CategoryEntity>().eq("parent_cid", parentCid));
+    }
+
 }
